* Move interactiveMode parsing into cliParser
